# src/assets/swagger.json
{
  "swagger": "2.0",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your JWT token in the format 'Bearer {token}'"
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": "Rest API for blockchain based health care system"
  },
  "host": "localhost:3000",
  "basePath": "/api/v1/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/users/": {
      "get": {
        "description": "This endpoint returns all the information about the user",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id_number": {
                  "type": "string",
                  "example": "123456",
                  "description": "The id number of the user",
                  "required": false,
                  "minLength": 4
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/hospitals/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Hospital"
        ]
      }
    },
    "/auth/register/user": {
      "post": {
        "description": "This endpoint registers a new user",
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id_number": {
                  "type": "string",
                  "example": "123456",
                  "description": "The id number of the user",
                  "required": true,
                  "minLength": 4
                },
                "name": {
                  "type": "string",
                  "example": "John Doe",
                  "description": "The name of the user",
                  "required": true,
                  "minLength": 4
                },
                "password": {
                  "type": "string",
                  "example": "123456",
                  "description": "The password of the user",
                  "required": true,
                  "minLength": 4
                },
                "insurance_entity": {
                  "type": "string",
                  "example": "123456",
                  "description": "The insurance entity of the user",
                  "required": true,
                  "minLength": 4
                },
                "blood_type": {
                  "type": "string",
                  "example": "A-",
                  "description": "The blood type of the user",
                  "required": true,
                  "maxLength": 3
                },
                "allergies": {
                  "type": "object",
                  "example": [
                    "allergy1",
                    "allergy2"
                  ],
                  "description": "The allergies of the user",
                  "required": true
                },
                "underlying_diseases": {
                  "type": "object",
                  "example": [
                    "disease1",
                    "disease2"
                  ],
                  "description": "The underlying diseases of the user",
                  "required": true
                },
                "fingerprint": {
                  "type": "string",
                  "example": "123456",
                  "description": "The fingerprint of the user",
                  "required": true,
                  "minLength": 4
                }
              }
            }
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/auth/register/hospital": {
      "post": {
        "description": "This endpoint registers a new hospital",
        "responses": {
          "201": {
            "description": "Created"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "parameters": [
          {
            "name": "hospital",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "John Doe",
                  "description": "The name of the hospital",
                  "required": true,
                  "minLength": 4
                },
                "username": {
                  "type": "string",
                  "example": "John Doe",
                  "description": "The username of the hospital",
                  "required": true,
                  "minLength": 4
                },
                "password": {
                  "type": "string",
                  "example": "123456",
                  "description": "The password of the hospital",
                  "required": true,
                  "minLength": 4
                }
              }
            }
          }
        ],
        "tags": [
          "Hospital"
        ]
      }
    },
    "/auth/login/user": {
      "post": {
        "description": "This endpoint logs in a user",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id_number": {
                  "type": "string",
                  "example": "123456",
                  "description": "The id number of the user",
                  "required": true,
                  "minLength": 4
                },
                "password": {
                  "type": "string",
                  "example": "123456",
                  "description": "The password of the user",
                  "required": true,
                  "minLength": 4
                }
              }
            }
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/auth/login/hospital": {
      "post": {
        "description": "This endpoint logs in a hospital",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "name": "hospital",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "John Doe",
                  "description": "The username of the hospital",
                  "required": true,
                  "minLength": 4
                },
                "password": {
                  "type": "string",
                  "example": "123456",
                  "description": "The password of the hospital",
                  "required": true,
                  "minLength": 4
                }
              }
            }
          }
        ],
        "tags": [
          "Hospital"
        ]
      }
    }
  }
}